requirements.txt
django
mongoengine
django-bootstrap-v5
ipykernel
python-dotenv
requests